fix:addressed maintainer feedback
